Drop compat code for `CartesianIndices` and `LinearIndices` from #446, #455, and #524
