+++ README.md
@@ -83,10 +83,6 @@ Currently, the `@compat` macro supports the following syntaxes:
 * `@compat finalizer(func, obj)` with the finalizer to run as the first argument and the object to be finalized
   as the second ([#24605]).
 
-* `CartesianIndices` and `LinearIndices` types represent cartesian and linear indices of
-  an array (respectively), and indexing such objects allows translating from one kind of index
-  to the other ([#25113]).
-
 * `codeunits(s)` returns an array-like view of the `UInt8` code units of
   a string and `ncodeunits(s)` returns the number of code units ([#25241]).
   `codeunit(s)` returns the type of the code units of `s` ([#24999]).
