+++ test/old.jl
@@ -613,3 +613,31 @@ end
 let A = [0, 0, 0], B = [1, 2, 3]
     @test unsafe_copyto!(A, 2, B, 1, 1) === A == [0, 1, 0]
 end
+
+# 0.7.0-DEV.3025
+let c = CartesianIndices((1:3, 1:2)), l = LinearIndices((1:3, 1:2))
+    @test LinearIndices(c) == collect(l)
+    @test CartesianIndices(l) == collect(c)
+    @test first(c) == CartesianIndex(1, 1)
+    @test CartesianIndex(1, 1) in c
+    @test first(l) == 1
+    @test size(c) == size(l) == (3, 2)
+    @test c == collect(c) == [CartesianIndex(1, 1) CartesianIndex(1, 2)
+                              CartesianIndex(2, 1) CartesianIndex(2, 2)
+                              CartesianIndex(3, 1) CartesianIndex(3, 2)]
+    @test l == collect(l) == reshape(1:6, 3, 2)
+    @test c[1:6] == vec(c)
+    @test l[1:6] == vec(l)
+    # TODO the following test fails on current Julia master (since 0.7.0-DEV.4742), and
+    # it's not clear yet whether it should work or not. See
+    # https://github.com/JuliaLang/julia/pull/26682#issuecomment-379762632 and the
+    # discussion following it
+    #@test l == l[c] == map(i -> l[i], c)
+    @test l[vec(c)] == collect(1:6)
+    @test CartesianIndex(1, 1) in CartesianIndices((3, 4))
+end
+
+# Issue #523
+@test length(Compat.CartesianIndices((1:1,))) == 1
+@test length(Compat.CartesianIndices((1:2,))) == 2
+@test length(Compat.CartesianIndices((1:2, -1:1))) == 6
